+++ mobile/library/cc/engine_builder.h
@@ -68,7 +68,7 @@ private:
   int dns_failure_refresh_seconds_max_ = 10;
   int dns_query_timeout_seconds_ = 25;
   std::string dns_preresolve_hostnames_ = "[]";
-  bool use_system_resolver_ = false;
+  bool use_system_resolver_ = true;
   int h2_connection_keepalive_idle_interval_milliseconds_ = 100000000;
   int h2_connection_keepalive_timeout_seconds_ = 10;
   int stats_flush_seconds_ = 60;